+++ src/test/kotlin/dev/ekvedaras/laravelquery/reference/ModelReferenceTest.kt
@@ -108,4 +108,28 @@ internal class ModelReferenceTest : BaseTestCase() {
         assertCompletion(*expected.toList().toTypedArray())
         assertNoCompletion(*notExpected.toList().toTypedArray())
     }
+
+    fun testResolveTableNameInsideScopeMethod() {
+        myFixture.configureByFile("model/modelInsideScope.php")
+
+        val table = DasUtil.getTables(db)
+            .filter { it.name == "users" }
+            .firstOrNull() ?: return fail("Did not find any tables.")
+        val columns = DasUtil.getColumns(table).map { it.name }
+        val otherTable = DasUtil.getTables(db)
+            .filterNot { it.name == "users" }
+            .lastOrNull() ?: return fail("Did not find any tables.")
+
+        val expected = schemas + columns
+        val notExpected =
+            schemaTables.entries.filterNot { it.key == table.dasParent?.name }.map { it.value }
+                .flatten() + // Tables of other schemas
+                DasUtil.getColumns(otherTable)
+                    .filterNot { columns.contains(it.name) }
+                    .map { it.name } // Columns of other table
+
+        myFixture.completeBasic()
+        assertCompletion(*expected.toList().toTypedArray())
+        assertNoCompletion(*notExpected.toList().toTypedArray())
+    }
 }