+++ src/test/resources/model/modelInsideScope.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App {
+class User extends \Illuminate\Database\Eloquent\Model
+{
+    public function scopeFirstId(\Illuminate\Database\Eloquent\Builder $query)
+    {
+        return $query->where('<caret>');
+    }
+}
+}