+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # Laravel Query Changelog
 
 ## [Unreleased]
+### Added
+- Resolve table name from model when inside model's scope method
 ### Fixed
 - Column reference resolving is now context aware, so it will not resolve to all columns of the same name if it knows the table.